fix: 이미지 Crop 방식 변경 및 파일화 로직 수정
