+++ app/src/main/AndroidManifest.xml
@@ -2,9 +2,9 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools">
 
-    <uses-feature
-        android:name="android.hardware.camera"
-        android:required="false" />
+
+    <uses-feature android:name="android.hardware.camera.any" />
+    <uses-permission android:name="android.permission.RECORD_AUDIO" />
 
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION" />
@@ -13,7 +13,8 @@
     <uses-permission android:name="android.permission.INTERNET" />
 
     <!--    갤러리 권한-->
-    <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" />
+    <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE"
+        android:maxSdkVersion="32" />
     <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE"
         android:maxSdkVersion="28" /> <!-- Android 10 이하에서만 유효 -->
 
@@ -35,6 +36,6 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Beginvegan"
         android:usesCleartextTraffic="true"
-        tools:targetApi="31"></application>
+        tools:targetApi="31" />
 
 </manifest>
