+++ presentation/src/main/AndroidManifest.xml
@@ -42,9 +42,11 @@
             android:exported="true"
             android:windowSoftInputMode="adjustPan" />
         <activity
-            android:name=".view.gallery.view.GalleryActivity"
+            android:name=".view.Image.gallery.view.GalleryActivity"
             android:exported="true"
             android:windowSoftInputMode="adjustResize" />
+        <activity android:name=".view.Image.camera.view.CameraActivity"
+            android:exported="true"/>
     </application>
 
 </manifest>
