+++ presentation/src/main/java/com/example/presentation/view/Image/gallery/model/GalleryImage.kt
@@ -1,0 +1,13 @@
+package com.example.presentation.view.Image.gallery.model
+
+import android.graphics.Bitmap
+import android.net.Uri
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+data class GalleryImage(
+    val imageUri: Uri?,
+    val check: Boolean = false,
+    val imagePath: String? = null
+) : Parcelable
