+++ presentation/src/main/java/com/example/presentation/view/login/viewModel/LoginViewModel.kt
@@ -1,5 +1,6 @@
 package com.example.presentation.view.login.viewModel
 
+import android.annotation.SuppressLint
 import android.content.Context
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
@@ -39,7 +40,9 @@ class LoginViewModel @Inject constructor(
         viewModelScope.launch {
             signInUseCase.invoke(email, providerId).onSuccess {
                 Timber.d("$it")
-                additionalInfoProvided = it.additionalInfo
+              /** 카메라 테스트를 위해 막아둔 거 나중에 지워야함 **/
+//                additionalInfoProvided = it.additionalInfo
+
                 _loginState.value = true
             }.onFailure {
                 _loginState.value = false
@@ -81,11 +84,7 @@ class LoginViewModel @Inject constructor(
                 Timber.d("KaKao User 사용자 정보 요청 실패 $error")
             } else if (user != null) {
                 Timber.d(
-                    "KaKao User 사용자 정보 요청 성공" +
-                            "\n회원번호: ${user.id}" +
-                            "\n이메일: ${user.kakaoAccount?.email}" +
-                            "\n닉네임: ${user.kakaoAccount?.profile?.nickname}" +
-                            "\n프로필사진: ${user.kakaoAccount?.profile?.thumbnailImageUrl}"
+                    "KaKao User 사용자 정보 요청 성공\n회원번호: ${user.id}\n이메일: ${user.kakaoAccount?.email}"
                 )
                 user.kakaoAccount?.email?.let { signIn(it,user.id.toString()) }
