+++ presentation/src/main/res/navigation/nav_gallery_graph.xml
@@ -6,7 +6,7 @@
 
     <fragment
         android:id="@+id/galleryListFragment"
-        android:name="com.example.presentation.view.gallery.view.GalleryListFragment"
+        android:name="com.example.presentation.view.Image.gallery.view.GalleryListFragment"
         android:label="GalleryListFragment" >
         <action
             android:id="@+id/action_galleryListFragment_to_selectImageFragment"
@@ -14,7 +14,7 @@
     </fragment>
     <fragment
         android:id="@+id/selectImageFragment"
-        android:name="com.example.presentation.view.gallery.view.SelectImageFragment"
+        android:name="com.example.presentation.view.Image.gallery.view.SelectImageFragment"
         android:label="SelectImageFragment" >
     </fragment>
 </navigation>
